% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addMetaData}
\alias{addMetaData}
\title{Annotate ctrdata function return values}
\usage{
addMetaData(x, con)
}
\arguments{
\item{x}{object to be annotated}

\item{con}{A \link[nodbi]{src} connection object, as obtained with
nodbi::\link[nodbi]{src_mongo}() or nodbi::\link[nodbi]{src_sqlite}()}
}
\description{
Annotate ctrdata function return values
}
\keyword{internal}
