% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkBinary}
\alias{checkBinary}
\title{checkBinary}
\usage{
checkBinary(b = NULL)
}
\arguments{
\item{b}{Vector of pre-defined binaries to be tested}
}
\value{
Logical, \code{TRUE} if all binaries ok
}
\description{
checkBinary
}
\keyword{internal}
