% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkDoc}
\alias{checkDoc}
\title{Check if a document exists based on its unique identier}
\usage{
checkDoc(con, id)
}
\arguments{
\item{con}{A \link[nodbi]{src} connection object, as obtained with
nodbi::\link[nodbi]{src_mongo}() or nodbi::\link[nodbi]{src_sqlite}()}
}
\value{
logical, FALSE (document or database does not
 exist) or TRUE
}
\description{
Check if a document exists based on its unique identier
}
\keyword{internal}
