% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{checkMongoVersionOk}
\alias{checkMongoVersionOk}
\title{Check the version of the build of the mongo server to be used}
\usage{
checkMongoVersionOk(mongo = rmongodb::mongo.create(host = "localhost:27017",
  db = "users"))
}
\arguments{
\item{mongo}{(\link{mongo}) A mongo connection object. If not provided, defaults to database "users" on localhost port 27017.}
}
\value{
A logical value indicating if the mongodb version is acceptable for use with this package.
}
\description{
In addition to the returned value, the function will generate a warning message if applicable.
}

