% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_path}
\alias{create_path}
\title{Create directory path recursively if it doesn't exist.}
\usage{
create_path(full_path)
}
\arguments{
\item{full_path}{Full path to create.}
}
\description{
Create directory path recursively if it doesn't exist.
}
