% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ctrCache}
\alias{ctrCache}
\title{Check, write, read cache object for ctrdata}
\usage{
ctrCache(xname, xvalue = NULL, verbose = TRUE)
}
\arguments{
\item{xname}{name of variable to read or write}

\item{xvalue}{value of variable to write}

\item{verbose}{set to `TRUE` to print debug info}
}
\value{
value of variable or `NULL` if variable does not exist
}
\description{
Check, write, read cache object for ctrdata
}
\keyword{internal}
