% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ctrConvertToJSON}
\alias{ctrConvertToJSON}
\title{ctrConvertToJSON}
\usage{
ctrConvertToJSON(tempDir, scriptName, verbose)
}
\arguments{
\item{tempDir}{Name of temporary directory with downloaded
trial information}

\item{scriptName}{Name of PHP or shell script to run}

\item{verbose}{Printing additional information if set to
\code{TRUE}; default is \code{FALSE}.}
}
\value{
System messages from converting
}
\description{
ctrConvertToJSON
}
\keyword{internal}
