% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrdata.R
\docType{package}
\name{ctrdata-package}
\alias{ctrdata-package}
\title{ctrdata: Retrieve and Analyze Clinical Trials Data from Public Registers}
\description{
A system for querying, retrieving and analyzing protocol- and results-related information on clinical trials from four public registers, the 'European Union Clinical Trials Register' ('EUCTR', \url{https://www.clinicaltrialsregister.eu/}), 'ClinicalTrials.gov' (\url{https://clinicaltrials.gov/} and also translating queries the retired classic interface), the 'ISRCTN' (\url{https://www.isrctn.com/}) and the 'European Union Clinical Trials Information System' ('CTIS', \url{https://euclinicaltrials.eu/}). Trial information is downloaded, converted and stored as JSON in a database ('PostgreSQL', 'SQLite', 'DuckDB', 'MongoDB' or 'MariaDB'; via package 'nodbi'). Protocols, statistical analysis plans, informed consent sheets and other documents in registers associated with trials can also be downloaded. Other functions implement trial analysis concepts canonically across registers, identify deduplicated records across registers, easily find and extract variables (fields) of interest even from complex nested data as used by registers, merge variables and update queries. The package can be used for monitoring, meta- and trend-analysis of the design and conduct as well as of the results of clinical trials across registers. See overview in Herold, R. (2025) \doi{10.1017/rsm.2025.10061}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cran.r-project.org/package=ctrdata}
  \item \url{https://rfhb.github.io/ctrdata/}
  \item Report bugs at \url{https://github.com/rfhb/ctrdata/issues}
}

}
\author{
\strong{Maintainer}: Ralf Herold \email{ralf.herold@mailbox.org} (\href{https://orcid.org/0000-0002-8148-6748}{ORCID})

Authors:
\itemize{
  \item Ralf Herold \email{ralf.herold@mailbox.org} (\href{https://orcid.org/0000-0002-8148-6748}{ORCID})
}

Other contributors:
\itemize{
  \item Marek Kubica (node-xml2js library) [copyright holder]
  \item Ivan Bozhanov (jstree library) [copyright holder]
}

}
\keyword{internal}
