% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrGenerateQueries.R
\name{ctrGenerateQueries}
\alias{ctrGenerateQueries}
\title{Generates queries that work across registers}
\usage{
ctrGenerateQueries(
  searchPhrase = NULL,
  condition = NULL,
  intervention = NULL,
  phase = NULL,
  population = NULL,
  recruitment = NULL,
  startBefore = NULL,
  startAfter = NULL,
  completedBefore = NULL,
  completedAfter = NULL,
  onlyMedIntervTrials = TRUE,
  onlyWithResults = FALSE,
  countries = NULL
)
}
\arguments{
\item{searchPhrase}{String with optional logical operators ("AND", "OR")
that will be searched in selected fields of registers (general or title
fields), should not include quotation marks}

\item{condition}{String with condition / disease}

\item{intervention}{String with intervention}

\item{phase}{String, e.g. "phase 2" (note that "phase 2+3" is a specific
category, not the union set of "phase 2" and "phase 3")}

\item{population}{String, e.g. "P" (paediatric), "A" (adult), "P+A"
(adult and paediatric), "E" (elderly), "P+A+E" participants can be recruited.
For ISRCTN, works only for "P", "A" or "E" but \emph{not mixed} populations.}

\item{recruitment}{String, one of "ongoing", "completed", "other"
(which includes "ended early" but this cannot be searched; use trial concept
\link{f.statusRecruitment} to identify this status)}

\item{startBefore}{String that can be interpreted as date (for EUCTR, when
trial was first registered)}

\item{startAfter}{String that can be interpreted as date (for EUCTR, when
trial was first registered)}

\item{completedBefore}{String that can be interpreted as date (does not work
with EUCTR)}

\item{completedAfter}{String that can be interpreted as date (does not work
with EUCTR)}

\item{onlyMedIntervTrials}{Logical, default \code{TRUE}, which indicates if
queries should search only for medicine interventional clinical trial}

\item{onlyWithResults}{Logical}

\item{countries}{Vector of country names, two- or three-letter ISO 3166 codes}
}
\value{
Named vector of URLs for finding trials in the registers and as
input to functions \link{ctrLoadQueryIntoDb},
\link{ctrOpenSearchPagesInBrowser} and \link[utils]{browseURL}
}
\description{
From high-level search terms provided by the user, generate queries that are
specific for each register with which \code{ctrdata} works,
see \link{ctrdata-registers}.
}
\details{
Search terms that are expanded to concepts such as from MeSH and MedDRA
by the search implementations in registers include the 'intervention' and
'condition' parameters. Logical operators only work in parameter 'searchPhrase'.
}
\examples{

ctrGenerateQueries(
  searchPhrase = "antibody AND covid",
  recruitment = "ongoing")

urls <- ctrGenerateQueries(
  intervention = "antibody",
  phase = "phase 3",
  population = "P+A",
  startAfter = "12/31/2020",
  completedBefore = "2025-01-01")

# open queries in register web interface
sapply(urls, utils::browseURL)

# For CTIS to accept such a search URL and show results,
# consider the extension, script and documentation at
# https://rfhb.github.io/ctrdata/index.html#
# id_2-script-to-automatically-copy-users-query-from-web-browser

# find potential research platform and platform trials
urls <- ctrGenerateQueries(
  searchPhrase = paste0(
   "basket OR platform OR umbrella OR master protocol OR ",
   "multiarm OR multistage OR subprotocol OR substudy OR ",
   "multi-arm OR multi-stage OR sub-protocol OR sub-study"),
 phase = "phase 3",
 startAfter = "01/31/2010",
 countries = c("DE", "US", "United Kingdom"))

\dontrun{
# count trials found
lapply(urls, ctrLoadQueryIntoDb, only.count = TRUE)

# load queries into database collection
dbc <- nodbi::src_sqlite(collection = "my_collection")
sapply(urls, ctrLoadQueryIntoDb, con = dbc, euctrprotocolsall = FALSE)
}

}
