% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ctrMongo}
\alias{ctrMongo}
\title{This is a function to set up connections to a Mongo DB server, one for
the actual trial records and a second for the keys as discovered by
variety in the trial records.}
\usage{
ctrMongo(collection = "ctrdata", db = "users",
  url = "mongodb://localhost", username = "", password = "",
  verbose = FALSE)
}
\arguments{
\item{collection}{name of collection}

\item{db}{name of database}

\item{url}{address of the mongodb server in mongo connection string
\href{{http://docs.mongodb.org/manual/reference/connection-string/}{URI format}}.}

\item{username}{In case access requires credentials.}

\item{password}{In case access requires credentials.}

\item{verbose}{emit some more output}
}
\value{
A mongo data base object, currently using mongolite
}
\description{
This is a function to set up connections to a Mongo DB server, one for
the actual trial records and a second for the keys as discovered by
variety in the trial records.
}
\keyword{internal}

