% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ctrRerunQuery}
\alias{ctrRerunQuery}
\title{ctrRerunQuery}
\usage{
ctrRerunQuery(querytoupdate = querytoupdate, debug = debug,
  collection = collection, uri = uri, password = password,
  verbose = verbose, queryupdateterm = queryupdateterm)
}
\arguments{
\item{querytoupdate}{Either the word "last" to re-run the last query that
was loaded into the collection, or the integer number of query to be run
again; see \link{dbQueryHistory}. This parameter takes precedence over
\code{queryterm}.}

\item{debug}{Printing additional information if set to \code{TRUE}; default
is \code{FALSE}.}

\item{collection}{Name of collection (default is "ctrdata")}

\item{uri}{Default is mongodb://localhost/users/.
Address of database in mongodb server, based on mongo connection string
format: mongodb://[username@]host1[:port1][,...hostN[:portN]]]/database/
Do NOT include password, this will only be used from the parameter.
See \url{http://docs.mongodb.org/manual/reference/connection-string/}}

\item{password}{In case access requires credentials.
Note this defaults to the environment variable "ctrdatamongopassword".
(by means of \code{Sys.getenv("ctrdatamongopassword")}), to
support scripting without revealing secrets.}

\item{verbose}{Print information.}
}
\description{
ctrRerunQuery
}
\keyword{internal}
