% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dbCTRAnnotateQueryRecords}
\alias{dbCTRAnnotateQueryRecords}
\title{dbQueryAnnotateRecords}
\usage{
dbCTRAnnotateQueryRecords(recordnumbers, annotations, annotation.text,
  annotation.mode, collection = collection, db = db, url = url,
  username = username, password = password, verbose = verbose)
}
\arguments{
\item{annotation.text}{Text to be including in the records retrieved
with the current query, in the field "annotation".}

\item{collection}{Name of collection (default is "ctrdata")}

\item{db}{Name of database (default is "users")}

\item{url}{Address of the mongodb server in mongo connection string URI format
\url{http://docs.mongodb.org/manual/reference/connection-string/} (default is
mongodb://localhost)}

\item{username}{In case access requires credentials.}

\item{password}{In case access requires credentials.}

\item{verbose}{Print information.}
}
\description{
dbQueryAnnotateRecords
}
\keyword{internal}
