% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{dbCTRGet}
\alias{dbCTRGet}
\title{Create a data frame from fields in the data base}
\usage{
dbCTRGet(fields = "")
}
\arguments{
\item{fields}{Vector of strings, with names of the sought fields. (Do not use a list)}
}
\value{
A data frame with columns corresponding to the sought fields.
}
\description{
Create a data frame from fields in the data base
}

