% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{dbCTRGetAll}
\alias{dbCTRGetAll}
\title{Create a data frame from all records in the data base}
\usage{
dbCTRGetAll(mongo = rmongodb::mongo.create(host = "localhost:27017", db =
  "users"), ns = "ctrdata")
}
\arguments{
\item{mongo}{(\link{mongo}) A mongo connection object. If not provided, defaults to database "users" on localhost port 27017.}

\item{ns}{Name of the collection in mongo database ("namespace"), defaults to "ctrdata"}
}
\value{
A data frame with columns corresponding to the sought fields.
}
\description{
Create a data frame from all records in the data base
}

