% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dbCTRUpdateQueryHistory}
\alias{dbCTRUpdateQueryHistory}
\title{dbCTRUpdateQueryHistory}
\usage{
dbCTRUpdateQueryHistory(register, queryterm, recordnumber,
  collection = collection, db = db, url = url, username = username,
  password = password, verbose = verbose, mongo = mongo)
}
\arguments{
\item{register}{Vector of abbreviations of registers to query, defaults to
"EUCTR"}

\item{queryterm}{Either a string with the full URL of a search in a register
or the data frame returned by the \link{ctrGetQueryUrlFromBrowser} or the
\link{dbQueryHistory} functions.
The queryterm is recorded in the collection \code{ns} for later use to
update records.}

\item{collection}{Name of collection (default is "ctrdata")}

\item{db}{Name of database (default is "users")}

\item{url}{Address of the mongodb server in mongo connection string URI format
\url{http://docs.mongodb.org/manual/reference/connection-string/} (default is
mongodb://localhost)}

\item{username}{In case access requires credentials.}

\item{password}{In case access requires credentials.}

\item{verbose}{Print information.}
}
\description{
dbCTRUpdateQueryHistory
}
\keyword{internal}
