% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetFieldsIntoDf.R
\name{dbGetFieldsIntoDf}
\alias{dbGetFieldsIntoDf}
\title{Create data frame of specified fields or trial concepts from database collection}
\usage{
dbGetFieldsIntoDf(fields = "", calculate = "", con, verbose = FALSE)
}
\arguments{
\item{fields}{Vector of one or more strings, with names of sought fields.
See function \link{dbFindFields} for how to find names of fields and
\link{ctrShowOneTrial} for interactively selecting field names.
Dot path notation ("field.subfield") without indices is supported.
If compatibility with \link[nodbi:src_postgres]{nodbi::src_postgres} is
needed, specify fewer than 50 fields, or use parent fields such as \code{"a.b"}
instead of \code{c("a.b.c.d", "a.b.c.e")} and then access sought fields with
\link{dfTrials2Long} followed by \link{dfName2Value} or with other R functions.}

\item{calculate}{Vector of one or more strings, which are names of functions
to calculate certain trial concepts from fields in the collection, across
different registers. See \link{ctrdata-trial-concepts} for available functions.}

\item{con}{A database connection object, created with
\code{nodbi}. See section \verb{1 - Database connection} in
\link{ctrdata}.}

\item{verbose}{If \code{TRUE}, prints additional information
(default \code{FALSE}).}
}
\value{
A data frame (or tibble, if \code{tibble} is loaded)
with columns corresponding to the sought fields.
A column with the record \verb{_id} will always be included.
The maximum number of rows of the returned data frame is equal to the number
of trial records in the database collection, or less if none of the fields
has a value in a record.
}
\description{
Fields in the collection are retrieved from all records into a data
frame (or tibble). The function uses the field names to appropriately type
the values that it returns, harmonising original values (e.g.,
"Yes" to \code{TRUE}, "false" to \code{FALSE}, "Information not present in EudraCT" to
\code{NA}, date strings to dates or time differences, number strings to numbers).
Trial concepts are calculated for all records and included in the return value.
}
\details{
Within a given trial record, a field can be hierarchical and structured,
that is, nested. The function simplifies the structure of nested data and
may concatenate multiple strings in a field using " / " (see example)
and may have widened the returned data frame with additional columns that
were recursively expanded from simply nested data (e.g., "externalRefs"
to columns "externalRefs.doi", "externalRefs.eudraCTNumber" etc.).
For an alternative ways for handling complex nested data, see
\link{dfTrials2Long} and \link{dfName2Value} for extracting
the sought variable(s).
}
\examples{

dbc <- nodbi::src_sqlite(
  dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
  collection = "my_trials",
  flags = RSQLite::SQLITE_RO)

# get fields that are nested within another field
# and can have multiple values with the nested field
dbGetFieldsIntoDf(
  fields = "b1_sponsor.b31_and_b32_status_of_the_sponsor",
  con = dbc)

# fields that are lists of string values are
# returned by concatenating values with " / "
dbGetFieldsIntoDf(
  fields = "keyword",
  con = dbc)

# calculate new field(s) from data across trials
df <- dbGetFieldsIntoDf(
  fields = "keyword",
  calculate = c("f.statusRecruitment", "f.isUniqueTrial", "f.startDate"),
  con = dbc)

table(df$.statusRecruitment, exclude = NULL)

\dontrun{
library(dplyr)
library(ggplot2)

df \%>\%
  filter(.isUniqueTrial) \%>\%
  count(.statusRecruitment)

df \%>\%
  filter(.isUniqueTrial) \%>\%
  ggplot() +
  stat_ecdf(aes(
    x = .startDate,
    colour = .statusRecruitment))
}

}
