% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfCalculate.R
\name{dfCalculate}
\alias{dfCalculate}
\title{Calculate fields from data in other fields}
\usage{
dfCalculate(name = ".*", df = NULL)
}
\arguments{
\item{name}{String with name of function to be applied to `df`, or regular
expression to list available functions.}

\item{df}{Optional. Data frame with fields needed to apply function `name`.}
}
\value{
Data frame with additional column of name `name` as calculated
applying the function `name`. If `df` is `NULL` (default), either a list of
functions corresponding to `name` or, if exactly one function is identified,
prints details of the function `name` to explain the implementation of the
concept and to return invisibly the names of fields needed for its calculation.
}
\description{
Across registers, calculate a trial concept (e.g., status of recruitment)
from fields in a data frame. The concept is defined within ctrdata
based on common understanding and any relevant publications.
Details of the function are printed to explain it, if no data frame is specified.
}
\details{
The names of the functions can also be used for the argument `calculate`
in \link{dbGetFieldsIntoDf} to achieve the same purpose already at the
time that a data frame with data is generated from the trial collection.
This approach will in most cases be more efficient and preferred over
using \link{dfCalculate}.
}
\examples{

# list names of all available functions in ctrdata
dfCalculate()

# list names of functions for regular expression of name
dfCalculate(name = "status")

# describe a specific function
dfCalculate(name = ".statusRecruitment")

# print descriptions of all functions
invisible(sapply(dfCalculate(), dfCalculate))

# apply dfCalculate to data frame
dbc <- nodbi::src_sqlite(
   dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
   collection = "my_trials",
   RSQLite::SQLITE_RO)

# use with existing data frame,
# first get fields needed
trialsDf <- dbGetFieldsIntoDf(
  fields = unlist(dfCalculate(name = ".statusRecruitment")),
  con = dbc)

# then calculate
dfCalculate(
  name = ".statusRecruitment",
  df = trialsDf)

# or use already when creating a trial data frame
# to subset to the unique records of clinical trials
trialsDf <- dbGetFieldsIntoDf(
  calculate = ".isUniqueTrial",
  con = dbc)
trialsDf[trialsDf[[".isUniqueTrial"]], ]


}
