% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dfMergeTwoVariablesRelevel}
\alias{dfMergeTwoVariablesRelevel}
\title{Merge two variables (superseded)}
\usage{
dfMergeTwoVariablesRelevel(df = NULL, colnames = "", levelslist = NULL)
}
\arguments{
\item{df}{A \link{data.frame} with the variables (columns) to be merged into
one vector.}

\item{colnames}{A vector of names of columns in `df` that hold the variables
to be merged, or a selection of columns as per \code{\link[dplyr]{select}}.}

\item{levelslist}{A names list with one slice each for a new value to be
used for a vector of old values (optional).}
}
\description{
Superseded, please use \link{dfMergeVariablesRelevel}
}
