% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_assignmentType.R
\name{f.assignmentType}
\alias{f.assignmentType}
\title{Calculate type of assignment to intervention in a study}
\usage{
f.assignmentType(df = NULL)
}
\arguments{
\item{df}{data frame such as from \link{dbGetFieldsIntoDf}. If \code{NULL},
prints fields needed in \code{df} for calculating this trial concept, which can
be used with \link{dbGetFieldsIntoDf}.}
}
\value{
data frame with columns \verb{_id} and \code{.assignmentType}, which is
a factor with levels \code{R} (randomised assignment) and \code{NR} (all other types
of assignment).
}
\description{
Calculate type of assignment to intervention in a study
}
\examples{
# fields needed
f.assignmentType()

# apply trial concept when creating data frame
dbc <- nodbi::src_sqlite(
  dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
  collection = "my_trials", flags = RSQLite::SQLITE_RO)
trialsDf <- dbGetFieldsIntoDf(
  field = "ctrname",
  calculate = "f.assignmentType",
  con = dbc)
trialsDf

}
