% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_isPlatformTrial.R
\name{f.isPlatformTrial}
\alias{f.isPlatformTrial}
\title{Calculate if study is likely a platform trial or not}
\usage{
f.isPlatformTrial(df = NULL)
}
\arguments{
\item{df}{data frame such as from \link{dbGetFieldsIntoDf}. If `NULL`,
prints fields needed in `df` for calculating this trial concept, which can
be used with \link{dfCalculateConcept}.}
}
\value{
data frame with columns `_id` and `.isPlatformTrial`, a logical.
}
\description{
Trial concept calculated: platform trial.
As operational definition, at least one of these criteria is true:
a. trial has "platform", "basket", "umbrella", "multi-?arm" or "multi-?stage"
in its title or description (for ISRCTN, this is the only criterion; some
trials in EUCTR lack data in English),
b. trial has more than 2 active arms with different investigational medicines,
after excluding comparator, auxiliary and placebo medicines (calculated with
function .numTestArmsSubstances(), not used for ISRCTN because it cannot be
calculated precisely),
c. trial more than 2 periods, after excluding safety run-in, screening,
enrolling, extension and follow-up periods (for CTGOV and CTGOV2, this
criterion requires results-related data).
Requires that EUCTR results have been included in the collection, using
ctrLoadQueryIntoDb(queryterm = ..., euctrresults = TRUE, con = ...).
Requires packages dplyr and stringdist to be installed; stringdist is used
for evaluating names of active substances, which are considered similar when
the similarity is 0.8 or higher.
}
\examples{
# fields needed
f.isPlatformTrial()

\dontrun{

# apply trial concept when creating data frame
dbc <- nodbi::src_sqlite(
  dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
  collection = "my_trials", flags = RSQLite::SQLITE_RO)
trialsDf <- dbGetFieldsIntoDf(
  calculate = "f.isPlatformTrial",
  con = dbc)
}


}
