% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_numSites.R
\name{f.numSites}
\alias{f.numSites}
\title{Calculate number of sites of a study}
\usage{
f.numSites(df = NULL)
}
\arguments{
\item{df}{data frame such as from \link{dbGetFieldsIntoDf}. If \code{NULL},
prints fields needed in \code{df} for calculating this trial concept, which can
be used with \link{dbGetFieldsIntoDf}.}
}
\value{
data frame with columns \verb{_id} and \code{.numSites}, an integer.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Trial concept calculated: number of the sites where the trial is conducted.
EUCTR lacks information on number of sites outside of the EEA;
for each non-EEA country mentioned, at least one site is assumed.
}
\examples{
# fields needed
f.numSites()

# apply trial concept when creating data frame
dbc <- nodbi::src_sqlite(
  dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
  collection = "my_trials", flags = RSQLite::SQLITE_RO)
trialsDf <- dbGetFieldsIntoDf(
  calculate = "f.numSites",
  con = dbc)
trialsDf

}
