% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_primaryEndpointResults.R
\name{f.primaryEndpointResults}
\alias{f.primaryEndpointResults}
\title{Calculate details of a study's primary endpoint statistical testing}
\usage{
f.primaryEndpointResults(df = NULL)
}
\arguments{
\item{df}{data frame such as from \link{dbGetFieldsIntoDf}. If \code{NULL},
prints fields needed in \code{df} for calculating this trial concept, which can
be used with \link{dbGetFieldsIntoDf}.}
}
\value{
data frame with columns \verb{_id} and new columns:
\code{.primaryEndpointFirstPvalue} (discarding any inequality indicator, e.g. <=),
\code{.primaryEndpointFirstPmethod} (normalised string, e.g. chisquared),
\code{.primaryEndpointFirstPsize} (number included in test, across assignment groups).
}
\description{
Trial concept calculated: Calculates several results-related elements of
the primary statistical analysis of the primary endpoint. Requires loading
results-related information.
For CTIS and ISRCTN, such information is not available in structured format.
Recommended to be combined with \link{f.controlType}, \link{f.sampleSize},
\link{f.assignmentType} and other \link{ctrdata-trial-concepts} for analyses.
}
\examples{
# fields needed
f.primaryEndpointResults()

# apply trial concept when creating data frame
dbc <- nodbi::src_sqlite(
  dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
  collection = "my_trials", flags = RSQLite::SQLITE_RO)
trialsDf <- dbGetFieldsIntoDf(
  calculate = "f.primaryEndpointResults",
  con = dbc)
trialsDf

}
