% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_sponsorType.R
\name{f.sponsorType}
\alias{f.sponsorType}
\title{Calculate type of sponsor of a study}
\usage{
f.sponsorType(df = NULL)
}
\arguments{
\item{df}{data frame such as from \link{dbGetFieldsIntoDf}. If \code{NULL},
prints fields needed in \code{df} for calculating this trial concept, which can
be used with \link{dbGetFieldsIntoDf}.}
}
\value{
data frame with columns \verb{_id} and \code{.sponsorType}, which is
a factor with levels \verb{for profit}, \verb{not for profit}, \code{mixed} (not and for
profit sponsors) or \code{other}.
}
\description{
Trial concept calculated: type or class of the sponsor(s) of the study.
No specific field is available in ISRCTN; thus, sponsor type is set to
\code{other}. Note: If several sponsors, sponsor type is deemed \code{mixed}
\emph{if there is both, a commercial and a non-commercial sponsor(s)}.
}
\examples{
# fields needed
f.sponsorType()

# apply trial concept when creating data frame
dbc <- nodbi::src_sqlite(
  dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
  collection = "my_trials", flags = RSQLite::SQLITE_RO)
trialsDf <- dbGetFieldsIntoDf(
  calculate = "f.sponsorType",
  con = dbc)
trialsDf

}
