% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_startDate.R
\name{f.startDate}
\alias{f.startDate}
\title{Calculate start date of a study}
\usage{
f.startDate(df = NULL)
}
\arguments{
\item{df}{data frame such as from \link{dbGetFieldsIntoDf}. If \code{NULL},
prints fields needed in \code{df} for calculating this trial concept, which can
be used with \link{dbGetFieldsIntoDf}.}
}
\value{
data frame with columns \verb{_id} and \code{.startDate}, a date.
}
\description{
Trial concept calculated: start of the trial, based on the
documented or planned start of recruitment, or on the date of opinion
of the competent authority.
}
\examples{
# fields needed
f.startDate()

# apply trial concept when creating data frame
dbc <- nodbi::src_sqlite(
  dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
  collection = "my_trials", flags = RSQLite::SQLITE_RO)
trialsDf <- dbGetFieldsIntoDf(
  field = "ctrname",
  calculate = "f.startDate",
  con = dbc)
trialsDf

}
