% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{findMongoimport}
\alias{findMongoimport}
\title{Convenience function to find location of mongoimport}
\usage{
findMongoimport()
}
\value{
Either an empty string if \code{mongoimport} was found on the path or, under MS Windows, a string representing the path
}
\description{
Convenience function to find location of mongoimport
}

