% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultseuctr.R
\name{format_euctr_results}
\alias{format_euctr_results}
\title{format_euctr_results}
\usage{
format_euctr_results(df, tabletype)
}
\arguments{
\item{df}{A data frame from a table node by rvest}

\item{tabletype}{Indicator if simple table (0)
or if it includes a subtable (higher than 0)}
}
\value{
A clean data frame
}
\description{
Function cleans tables from EUCTR results,
in particular by removing extraneous cells
and interpreting the visual / positional
markup used by EUCTR
}
