% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultseuctr.R
\name{getEuctrResultsTrial}
\alias{getEuctrResultsTrial}
\title{getEuctrResultsTrial}
\usage{
getEuctrResultsTrial(fullurleuctrtrialresults = "", eudract_number = "")
}
\arguments{
\item{fullurleuctrtrialresults}{A full URL to the web pages with the results
of a specific trial in the EUCTR, example
https://www.clinicaltrialsregister.eu/ctr-search/trial/2007-000371-42/results}

\item{eudract_number}{A single EudraCT number as string. If specified, takes
precedence over specifying fullurleuctrtrialresults.}
}
\value{
A list with named data frames representing the results.
}
\description{
Obtain the results published for a single trial in the EUCTR.
}
