% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{installCygwin}
\alias{installCygwin}
\title{Convenience function to install a cygwin environment under MS Windows, including perl}
\usage{
installCygwin(overwrite = FALSE)
}
\arguments{
\item{overwrite}{Set to true to force updating and overwriting an existing installation in \code{c:\\cygwin}}
}
\description{
Convenience function to install a cygwin environment under MS Windows, including perl
}

