% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{installCygwinWindowsTest}
\alias{installCygwinWindowsTest}
\title{Convenience function to test for working cygwin installation}
\usage{
installCygwinWindowsTest(verbose = FALSE)
}
\arguments{
\item{verbose}{If \code{TRUE}, prints confirmatory
message (default \code{FALSE})}
}
\value{
Information if cygwin can be used, \code{TRUE}
 or \code{FALSE}, or NULL if not under MS Windows
}
\description{
Convenience function to test for working cygwin installation
}
\keyword{internal}
