% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{installFindBinary}
\alias{installFindBinary}
\title{Check availability of binaries installed in operating system}
\usage{
installFindBinary(commandtest = NULL, debug = FALSE)
}
\arguments{
\item{commandtest}{Command to be used for testing the availability of the binary, e.g. "php -v". Note
internal quotes need to be escaped, e.g. \code{installFindBinary('php -r \"simplexml_load_string(\'\');\"')}.
See R/onload.R for tested binaries.}

\item{debug}{Set to \code{TRUE} to see printed return value of \code{commandtest}}
}
\value{
A logical if executing commandtest returned an error or not
}
\description{
Check availability of binaries installed in operating system
}
\keyword{internal}
