% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{installMongoCheckVersion}
\alias{installMongoCheckVersion}
\title{Check the version of the build of the mongo server to be used}
\usage{
installMongoCheckVersion(collection = "ctrdata", db = "users",
  url = "mongodb://localhost", username = "", password = "",
  verbose = FALSE)
}
\arguments{
\item{collection}{Name of collection (default is "ctrdata")}

\item{db}{Name of database (default is "users")}

\item{url}{Address of the mongodb server in mongo connection string URI format
\url{http://docs.mongodb.org/manual/reference/connection-string/} (default is
mongodb://localhost)}

\item{username}{In case access requires credentials.}

\item{password}{In case access requires credentials.}

\item{verbose}{Print information.}
}
\value{
A logical value indicating if the mongodb version is acceptable for
  use with this package.
}
\description{
In addition to the returned value, the function will generate a warning
message if applicable.
}
\keyword{internal}

