% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{installMongoFindBinaries}
\alias{installMongoFindBinaries}
\title{Detect location of local mongo database binaries (mongo,
mongoimport) and save into private environment}
\usage{
installMongoFindBinaries(mongoDir = NULL, debug = FALSE)
}
\arguments{
\item{mongoDir}{Vector of folders to search for mongo binaries, appended to defaults that include
"C:/Program Files/MongoDB/Server/3.6/binW, "c:/mongodb/bin", "/usr/local/opt/mongodb/bin" and "/usr/bin".}

\item{debug}{Printing additional information if set to \code{TRUE}; default
is \code{FALSE}.}
}
\value{
A vector of length 2 with the paths for the \code{mongo} and
  \code{mongoimport} binaries on the user's system
}
\description{
Call this function with a correct parameter if the automatic
detection fails.
}
\keyword{internal}
