% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabeling_values.R
\name{label_data}
\alias{label_data}
\title{Relabel values in CTR dataframe using a list of value-label pairs.}
\usage{
label_data(df, choices_list, lang = c("french", "english"))
}
\arguments{
\item{df}{A data.frame.}

\item{choices_list}{A named list of value-label pair data.frames.}

\item{lang}{Language column to use for relabeling. Default is both French and English.}
}
\value{
Data frame with relabeled values.
}
\description{
Relabel values in CTR dataframe using a list of value-label pairs.
}
