% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{mongo2df}
\alias{mongo2df}
\title{Convert a mongo query result object into a data frame}
\usage{
mongo2df(x)
}
\arguments{
\item{x}{A result of object of a mongo query}
}
\value{
A data frame with the data that was stored in the input object
}
\description{
Convert a mongo query result object into a data frame
}

