% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{openCTRWebBrowser}
\alias{openCTRWebBrowser}
\title{Open advanced search pages of register(s) in default web browser.}
\usage{
openCTRWebBrowser(register = c("EUCTR", "CTGOV"), copyright = FALSE, ...)
}
\arguments{
\item{register}{Register(s) to open. Either "EUCTR" or "CTGOV" or a vector of both. Default is to open both registers' advanced search pages.}

\item{copyright}{If set to \code{TRUE}, opens copyright pages of register.}

\item{...}{Any additional parameter to use with browseURL, which is called by this function.}
}
\value{
Is always true, invisibly.
}
\description{
Open advanced search pages of register(s) in default web browser.
}

