% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_save}
\alias{split_save}
\title{Split data by a month column and save anonymized data.}
\usage{
split_save(df, month_var = "year_mon", path = box_data_path)
}
\arguments{
\item{df}{Data frame to split and save.}

\item{month_var}{Column name representing year-month.}

\item{path}{Base path to save data.}
}
\description{
Split data by a month column and save anonymized data.
}
