% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultseuctr.R
\name{strnorm}
\alias{strnorm}
\title{strnorm}
\usage{
strnorm(s)
}
\arguments{
\item{s}{A string to be normalised}
}
\value{
Normalised string
}
\description{
This function normalises a single string
according to some rules developed from
processing EUCTR result field names
}
