% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{temp_loader}
\alias{temp_loader}
\title{Load Excel file through a temporary path to handle permissions or long paths.}
\usage{
temp_loader(path, ...)
}
\arguments{
\item{path}{Path to Excel file.}

\item{...}{Additional arguments to pass to read_excel.}
}
\value{
A data.frame.
}
\description{
Load Excel file through a temporary path to handle permissions or long paths.
}
