% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{temp_uploader}
\alias{temp_uploader}
\title{Upload a data.frame to a specified path using a temporary file.}
\usage{
temp_uploader(df, path)
}
\arguments{
\item{df}{The data frame to save.}

\item{path}{Destination file path.}
}
\description{
Upload a data.frame to a specified path using a temporary file.
}
