% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{testCygwin}
\alias{testCygwin}
\title{Convenience function to test for working cygwin installation}
\usage{
testCygwin()
}
\value{
Information if cygwin can be used, \code{TRUE} or \code{FALSE}
}
\description{
Convenience function to test for working cygwin installation
}

