% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{typeField}
\alias{typeField}
\title{Change type of field based on name of field}
\usage{
typeField(dfi)
}
\arguments{
\item{dfi}{a data frame of columns _id, fieldname}
}
\description{
Change type of field based on name of field
}
\keyword{internal}
