% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addMetaData}
\alias{addMetaData}
\title{Annotate ctrdata function return values}
\usage{
addMetaData(x, uri, collection, password)
}
\arguments{
\item{x}{object to be annotated}

\item{uri}{Default is mongodb://localhost/users/.
Address of database in mongodb server, based on mongo connection string
format: mongodb://[username@]host1[:port1][,...hostN[:portN]]]/database/
Do NOT include password, this will only be used from the parameter.
See \url{http://docs.mongodb.org/manual/reference/connection-string/}}

\item{collection}{Name of collection (default is "ctrdata")}

\item{password}{In case access requires credentials.
Note this defaults to the environment variable "ctrdatamongopassword".
(by means of \code{Sys.getenv("ctrdatamongopassword")}), to
support scripting without revealing secrets.}
}
\description{
Annotate ctrdata function return values
}
\keyword{internal}
