% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addMetaData}
\alias{addMetaData}
\title{Title}
\usage{
addMetaData(x, url, db, collection, username, password)
}
\arguments{
\item{x}{object to be annotated}

\item{url}{Address of the mongodb server in mongo connection string URI format
\url{http://docs.mongodb.org/manual/reference/connection-string/} (default is
mongodb://localhost)}

\item{db}{Name of database (default is "users")}

\item{collection}{Name of collection (default is "ctrdata")}

\item{username}{In case access requires credentials.}

\item{password}{In case access requires credentials.}
}
\description{
Title
}
\keyword{internal}
