% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctr_data_loader.R
\name{ctr_data_loader}
\alias{ctr_data_loader}
\title{Load CTR data between two dates.}
\usage{
ctr_data_loader(path = ".", start_date = NULL, end_date = NULL)
}
\arguments{
\item{path}{Root path where data is stored.}

\item{start_date}{Start date in yyyy-mm-dd format.}

\item{end_date}{End date in yyyy-mm-dd format.}
}
\value{
A data frame of records between specified dates.
}
\description{
Load CTR data between two dates.
}
\examples{
box_path = "~/Box/DMAC/Data Anonymized/"

df <- ctr_data_loader(
  path = box_path,
  start_date = "2025-05-01",
  end_date = "2025-05-31"
)
head(tibble::as_tibble(df))

}
