% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrdata-registers.R
\docType{data}
\name{ctrdata-registers}
\alias{ctrdata-registers}
\title{Information on clinical trial registers}
\description{
Registers of the four clinical trial registers from which package
\link{ctrdata} can retrieve, aggregate and analyse protocol- and
result-related information as well as documents, last updated 2025-07-24.
}
\section{1 - Overview}{

\itemize{
\item \strong{EUCTR}: The EU Clinical Trials Register holds more than 44,350 clinical
trials (at least one investigational medicinal product, IMP; in the European
Union and beyond), including more than 25,400 trials with results, which
continue to be added (and can be loaded by \code{ctrdata}).
\item \strong{CTIS}: The EU Clinical Trials Information System, launched in 2023,
holds about 9,800 publicly accessible clinical trials, including
more than 270 with results or a report (only as PDF files).
\emph{No results in a structured electronic format are foreseeably available},
thus \code{ctrdata} cannot load any CTIS results.
(To automatically get CTIS search query URLs, see
\ifelse{latex}{\out{\href{https://rfhb.github.io/ctrdata/\#id_2-script-to-automatically-copy-users-query-from-web-browser}{here}}}{\href{https://rfhb.github.io/ctrdata/#id_2-script-to-automatically-copy-users-query-from-web-browser}{here}})
\item \strong{CTGOV2}: ClinicalTrials.gov holds more than 550,000 interventional and
observational studies, including almost 70,000 interventional studies with
results (can be loaded by \code{ctrdata}).
\item \strong{ISRCTN}: The ISRCTN Registry holds almost 26,800 interventional and
observational health studies, including more than 14,600 studies with
results (as references).
\emph{No results in a structured electronic format are foreseeably available},
thus \code{ctrdata} cannot load any ISRCTN results.
}
}

\section{2 - Notable changes}{


CTGOV "classic" was retired on 2024-06-25; \code{ctrdata} subsequently translates
CTGOV queries to CTGOV2 queries. The new website ("CTGOV2") can be used with
\code{ctrdata} since 2023-08-27. Database collections created with CTGOV queries
can still be used since functions in \code{ctrdata} continue to support them.
CTIS was relaunched on 2024-06-17, changing the data structure and search
syntax, to which \code{ctrdata} was updated.
CTIS can be used with \code{ctrdata} since 2023-03-25.
EUCTR removed search parameter \verb{status=} as of February 2025.
More information on changes:
\href{https://rfhb.github.io/ctrdata/news/index.html}{here}.
}

\section{3 - References}{
\tabular{lcccc}{
   \strong{Material} \tab \strong{EUCTR} \tab \strong{CTGOV2} \tab \strong{ISRCTN} \tab \strong{CTIS} \cr
   About \tab \href{https://www.clinicaltrialsregister.eu/about.html}{link} \tab \href{https://clinicaltrials.gov/about-site/about-ctg}{link} \tab \href{https://www.isrctn.com/page/about}{link} \tab \href{https://euclinicaltrials.eu/about-this-website/}{link} \cr
   Terms & conditions, disclaimer \tab \href{https://www.clinicaltrialsregister.eu/disclaimer.html}{link} \tab \href{https://clinicaltrials.gov/about-site/terms-conditions}{link} \tab \ifelse{latex}{\out{\href{https://www.isrctn.com/page/faqs\#using-the-isrctn}{link}}}{\href{https://www.isrctn.com/page/faqs#using-the-isrctn}{link}} \tab \href{https://euclinicaltrials.eu/guidance-and-q-as/}{link} \cr
   How to search \tab \href{https://www.clinicaltrialsregister.eu/doc/How_to_Search_EU_CTR.pdf}{link} \tab \href{https://clinicaltrials.gov/find-studies/how-to-search}{link} \tab \href{https://www.isrctn.com/page/search-tips}{link} \tab \href{https://euclinicaltrials.eu/search-tips-and-guidance/}{link} \cr
   Search interface \tab \href{https://www.clinicaltrialsregister.eu/ctr-search/search}{link} \tab \href{https://clinicaltrials.gov/}{link} \tab \href{https://www.isrctn.com/}{link} \tab \href{https://euclinicaltrials.eu/search-for-clinical-trials/}{link} \cr
   Expert / advanced search \tab \href{https://www.clinicaltrialsregister.eu/ctr-search/search}{link} \tab \href{https://clinicaltrials.gov/expert-search}{link} \tab \href{https://www.isrctn.com/editAdvancedSearch}{link} \tab \href{https://euclinicaltrials.eu/ctis-public/search}{link} \cr
   Glossary / related information \tab \href{https://www.clinicaltrialsregister.eu/doc/EU_Clinical_Trials_Register_Glossary.pdf}{link} \tab \href{https://clinicaltrials.gov/study-basics/glossary}{link} \tab \href{https://www.who.int/clinical-trials-registry-platform/network/who-data-set}{link} \tab \href{https://accelerating-clinical-trials.europa.eu/}{link} \cr
   FAQ / caveats / examples \tab \href{https://www.clinicaltrialsregister.eu/doc/EU_CTR_FAQ.pdf}{link} \tab \href{https://clinicaltrials.gov/policy/faq}{link}, \href{https://clinicaltrials.gov/about-site/selected-publications}{link}, \href{https://clinicaltrials.gov/submit-studies/prs-help/support-training-materials#example-studies}{link} \tab \href{https://www.isrctn.com/page/faqs}{link} \tab \href{https://euclinicaltrials.eu/website-outages-and-system-releases/}{link} \cr
   Data dictionaries / structure \tab \href{https://eudract.ema.europa.eu/result.html}{link}, \href{https://eudract.ema.europa.eu/docs/technical/EudraCT\%20protocol\%20related\%20data\%20dictionary.xls}{link}, \href{https://eudract.ema.europa.eu/docs/technical/V7_V8_Country_List_20210804.xlsx}{link} \tab \href{https://cdn.clinicaltrials.gov/documents/xsd/public.xsd}{link}, \href{https://clinicaltrials.gov/data-about-studies/study-data-structure}{link}, \href{https://cdn.clinicaltrials.gov/documents/tutorial/content/index.html}{link} \tab \href{https://www.isrctn.com/page/definitions}{link} \tab \href{https://www.ema.europa.eu/en/human-regulatory-overview/research-development/clinical-trials-human-medicines/clinical-trials-information-system-ctis-training-support}{link} (see XLSX files) \cr
   Example* (see below) \tab \ifelse{latex}{\out{\href{https://www.clinicaltrialsregister.eu/ctr-search/search?query=Infections\&age=newborn\&age=preterm-new-born-infants}{link}}}{\href{https://www.clinicaltrialsregister.eu/ctr-search/search?query=Infections&age=newborn&age=preterm-new-born-infants}{link}} \tab \ifelse{latex}{\out{\href{https://clinicaltrials.gov/search?ageRange=0M_1M\&cond=Infections\&aggFilters=studyType:int\&distance=50\&intr=Investigational+Agent}{link}}}{\href{https://clinicaltrials.gov/search?ageRange=0M_1M&cond=Infections&aggFilters=studyType:int&distance=50&intr=Investigational+Agent}{link}} \tab \ifelse{latex}{\out{\href{https://www.isrctn.com/search?q=\&filters=ageRange:Neonate,conditionCategory:Infections+and+Infestations\&searchType=advanced-search}{link}}}{\href{https://www.isrctn.com/search?q=&filters=ageRange:Neonate,conditionCategory:Infections+and+Infestations&searchType=advanced-search}{link}} \tab \ifelse{latex}{\out{\href{https://euclinicaltrials.eu/ctis-public/search\#searchCriteria={"containAll":"infection","containAny":"neonates","containNot":""}}{link}}}{\href{https://euclinicaltrials.eu/ctis-public/search#searchCriteria={"containAll":"infection","containAny":"neonates"}}{link}} \cr
}


Some registers are expanding entered search terms using dictionaries
(\href{https://clinicaltrials.gov/data-api/about-api/search-areas}{example}).
}

\section{4 - Example and ctrdata motivation}{


See \code{vignette("ctrdata_summarise")} for several other examples.

*This example is an expert search for interventional trials primarily with
neonates, investigating treatments for infectious conditions. It shows that
searches in the web interface of most registers are not sufficient to
identify the trials of interest:
\itemize{
\item EUCTR retrieves trials with neonates, but not only those exclusively in neonates.
\item ISRCTN retrieves studies with interventions other than medicines.
\item CTIS retrieves trials that mention the words neonates and infection. (To
show CTIS search results, see \ifelse{latex}{\out{\href{https://rfhb.github.io/ctrdata/\#id_2-script-to-automatically-copy-users-query-from-web-browser}{here}}}{\href{https://rfhb.github.io/ctrdata/#id_2-script-to-automatically-copy-users-query-from-web-browser}{here}})
}

To address this issue, trials can be retrieved with \link{ctrLoadQueryIntoDb}
into a database collection and in a second step trials of interest can be
selected based on values of relevant fields, for example:
\itemize{
\item EUCTR field \code{f115_children_211years} and other age group criteria
\item ISRCTN field \code{interventions.intervention.interventionType} for type of study
\item CTIS fields \code{ageGroup} and \code{authorizedApplication.authorizedPartI.medicalConditions.medicalCondition}
}

\code{ctrdata} supports users with pre-defined \link{ctrdata-trial-concepts} and
these cover the example above, and with functions \link{dbFindFields} and
\link{ctrShowOneTrial} for finding fields of interest and reviewing data
structure, respectively.
}

\author{
Ralf Herold \email{ralf.herold@mailbox.org}
}
\keyword{data}
