% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrdata-trial-concepts.R
\docType{data}
\name{ctrdata-trial-concepts}
\alias{ctrdata-trial-concepts}
\title{Trial concepts implemented across registers}
\description{
\code{ctrdata} includes (since version 1.21.0) functions that implement selected
trial concepts. Concepts of clinical trials, such as their start or status of
recruitment, require to analyse several fields against various
pre-defined values. In addition, the structure and value sets of data
differ between all \link{ctrdata-registers}.
The included trial concepts are intended to simplify and accelerate a
user's workflow as well as to provide consistency of analyses.
}
\details{
The implementation of trial concepts in \code{ctrdata} is based on current
understanding, publication of data models and scientific papers as relevant.
Call \link{dfCalculate} for one or more such functions to review the
explanation and the fields needed. As with other \code{R} functions,
the code is printed when entering as command just the name of the function,
e.g. \code{.startDate}.
Raise an issue \href{https://github.com/rfhb/ctrdata/issues}{here}
to ask about or improve a concept.

The following trial concepts can be used by referencing their name when
calling \link{dbGetFieldsIntoDf} (parameter \code{calculate}) and
\link{dfCalculate} (parameter \code{name}).
Concepts will continue to be refined and added;
last updated 2025-02-09.
\itemize{
\item \strong{.controlType} (factor) which type of internal or concurrent control is
used in the trial? ("none", "no-treatment", "placebo", "active",
"placebo+active" or "other")
\item \strong{.isMedIntervTrial} (logical) is the trial interventional and does
it have one or more medicines (drugs or biological) as investigational
(experimental) intervention? (irrespective of status of authorisation and
of study design)
\item \strong{.isPlatformTrial} (logical) is the trial possibly a (research) platform
trial? (based on title, .numTestArmsSubstances, number of periods)
\item \strong{.isUniqueTrial} (logical) is the trial record unique in the data frame
of trial, based on the default parameters of \link{dbFindIdsUniqueTrials}?
\item \strong{.numTestArmsSubstances} (integer) how many arms or groups have
medicines that are investigational? (cannot be calculated for ISRCTN)
\item \strong{.startDate} (date) the planned, authorised or documented date
of start of recruitment
\item \strong{.sampleSize} (integer) the achieved (or planned) number of subjects or
participants recruited
\item \strong{.statusRecruitment} (factor) a status that is simplified to
ongoing (includes temporarily halted), completed and other (includes
planned, stopped, terminated)
\item \strong{.trialPhase} (ordered factor) the phase(s) of medicine development
with which a trial is associated
}
}
\author{
Ralf Herold \email{ralf.herold@mailbox.org}
}
\keyword{data}
