% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrdata.R
\docType{package}
\name{ctrdata}
\alias{ctrdata}
\alias{ctrdata-package}
\title{ctrdata: A package to retrieve and to prepare for analysis, information on
clinical trials from public registers.}
\description{
The ctrdata package provides three categories of important functions:
operations on a clinical trial register (ctrOpenSearchPagesInBrowser,
ctrGetQueryUrlFromBrowser, ctrQueryHistoryInDb, ctrLoadQueryIntoDb), on the
database that holds downloaded information (dbGetVariablesIntoDf,
dbFindVariable, dbFindIdsUniqueTrials) and on R dataframes with clinical
trial information (dfMergeTwoVariablesRelevel, dfMergeTwoVariablesRelevel).
}
\section{ctrdata functions}{
 The foo functions ...
}

