% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ctrDb}
\alias{ctrDb}
\title{Check and prepare nodbi connection object for ctrdata}
\usage{
ctrDb(con = nodbi::src_sqlite(collection = "ctrdata_auto_generated"))
}
\arguments{
\item{con}{A connection object, see section
`Databases` in \link{ctrdata-package}}
}
\value{
Connection object as list, with collection
 element under root
}
\description{
Check and prepare nodbi connection object for ctrdata
}
\keyword{internal}
