% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ctrDb}
\alias{ctrDb}
\title{Check and prepare nodbi connection object for ctrdata}
\usage{
ctrDb(con = nodbi::src_sqlite(collection = "ctrdatagenerated"))
}
\arguments{
\item{con}{A \link[nodbi]{src} connection object, as obtained with
nodbi::\link[nodbi]{src_mongo}() or nodbi::\link[nodbi]{src_sqlite}()}
}
\value{
Connection object as list, with collection
 element under root
}
\description{
Check and prepare nodbi connection object for ctrdata
}
\keyword{internal}
