% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{ctrDb}
\alias{ctrDb}
\title{Check and prepare nodbi connection object for ctrdata}
\usage{
ctrDb(con)
}
\arguments{
\item{con}{A database connection object, created with
\code{nodbi}. See section \verb{1 - Database connection} in
\link{ctrdata}.}
}
\value{
Connection object as list, with collection
element under root
}
\description{
Check and prepare nodbi connection object for ctrdata
}
\keyword{internal}
