% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrFindActiveSubstanceSynonyms.R
\name{ctrFindActiveSubstanceSynonyms}
\alias{ctrFindActiveSubstanceSynonyms}
\title{Find synonyms of an active substance}
\usage{
ctrFindActiveSubstanceSynonyms(activesubstance = "", verbose = FALSE)
}
\arguments{
\item{activesubstance}{An active substance, in an atomic character vector}

\item{verbose}{Print number of studies found in CTGOV2 for `activesubstance`}
}
\value{
A named character vector of the active substance (input parameter),
 the MeSH code(s) and various names used in registered studies, or NULL if
 active substance was not found and may be invalid. The active substances
 are ordered in decreasing number of occurrence.
}
\description{
An active substance can be identified by a recommended international
nonproprietary name (INN), a trade or product name, or a company code(s).
To find likely synonyms, the function retrieves from CTGOV2 the field
protocolSection.armsInterventionsModule.interventions.
Note this is mostly manually filled, thus may not be free of errors.
}
\examples{
\dontrun{

ctrFindActiveSubstanceSynonyms(activesubstance = "imatinib")
activesubstance                mesh
     "imatinib" "imatinib mesylate"  "imatinib"  "gleevec"  "imatinib mesylate"
       "glivec"            "STI571"    "111201" "CGP57148"          "CGP57148B"
      "gleevac" "mesylate imatinib"    "ST1571"
}

}
