% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ctrFindActiveSubstanceSynonyms}
\alias{ctrFindActiveSubstanceSynonyms}
\title{Find synonyms of an active substance}
\usage{
ctrFindActiveSubstanceSynonyms(activesubstance = "")
}
\arguments{
\item{activesubstance}{An active substance, in an atomic character vector}
}
\value{
A character vector of the active substance (input parameter) and
 synonyms, if any were found
}
\description{
An active substance can be identified by a recommended international
nonproprietary name, a trade or product name, or a company code(s).
}
\details{
At this time, this function uses the register ClinicalTrials.Gov to
detect which substances were also searched for.
}
\examples{

\dontrun{
ctrFindActiveSubstanceSynonyms(
  activesubstance = "imatinib"
)
}

}
