% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrGenerateQueries.R
\name{ctrGenerateQueries}
\alias{ctrGenerateQueries}
\title{Generates queries that work across registers}
\usage{
ctrGenerateQueries(
  condition = NULL,
  intervention = NULL,
  recruitment = NULL,
  phase = NULL,
  startBefore = NULL,
  startAfter = NULL,
  completedBefore = NULL,
  completedAfter = NULL,
  onlyWithResults = FALSE,
  registers = c("EUCTR", "ISRCTN", "CTIS", "CTGOV2")
)
}
\arguments{
\item{condition}{String with condition / disease}

\item{intervention}{String with intervention}

\item{recruitment}{String, one of "ongoing", "completed", "other"}

\item{phase}{String, e.g. "phase 2" (note that "phase 2+3" is a specific
category, not the union set of "phase 2" and "phase 3")}

\item{startBefore}{String that can be interpreted as date, see example}

\item{startAfter}{String that can be interpreted as date}

\item{completedBefore}{String that can be interpreted as date (does not work
with EUCTR)}

\item{completedAfter}{String that can be interpreted as date (does not work
with EUCTR)}

\item{onlyWithResults}{Logical}

\item{registers}{Vector of register names, default all registers}
}
\value{
Named vector of URLs for finding trials in the registers and as
input to functions \link{ctrLoadQueryIntoDb} and
\link{ctrOpenSearchPagesInBrowser}
}
\description{
From high-level search terms provided by the user, generate specific queries
for each registers with which ctrdata works, see \link{ctrdata-registers}.
Search terms that are expanded to concepts such as from MeSH and MedDRA
by the search implementations in registers include the
intervention and condition, see \link{ctrdata-registers}.
}
\examples{

urls <- ctrGenerateQueries(
  intervention = "antibody",
  recruitment = "ongoing",
  phase = "phase 3",
  startAfter = "2000-01-01")

# open queries in register web interface
sapply(urls, ctrOpenSearchPagesInBrowser)

# load queries into database collection
# sapply(urls, ctrLoadQueryIntoDb, con = dbc)

}
