% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ctrGetQueryUrlFromBrowser}
\alias{ctrGetQueryUrlFromBrowser}
\title{Import from clipboard the URL of a search in one of the registers}
\usage{
ctrGetQueryUrlFromBrowser(content = "")
}
\arguments{
\item{content}{URL from browser address bar.
Defaults to clipboard contents.}
}
\value{
A string of query parameters that can be used to retrieve data
from the register.

A data frame with a query term and the register name that can
directly be used in \link{ctrLoadQueryIntoDb} and in
\link{ctrOpenSearchPagesInBrowser}
}
\description{
Import from clipboard the URL of a search in one of the registers
}
\examples{

\dontrun{
db <- nodbi::src_sqlite(collection = "test")
ctrLoadQueryIntoDb(ctrGetQueryUrlFromBrowser(), con = db)
}

}
