% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ctrGetQueryUrlFromBrowser}
\alias{ctrGetQueryUrlFromBrowser}
\title{Import from clipboard the URL of a search in one of the registers}
\usage{
ctrGetQueryUrlFromBrowser(content = clipr::read_clip())
}
\arguments{
\item{content}{URL from browser address bar. Defaults to clipboard contents.}
}
\value{
A string of query parameters that can be used to retrieve data from the register.

A list with a query term and the register name that can directly be used in \link{ctrLoadQueryIntoDb}
}
\description{
Import from clipboard the URL of a search in one of the registers
}
\examples{
\dontrun{
ctrLoadQueryIntoDb (ctrGetQueryUrlFromBrowser())
}

}

