% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ctrLoadQueryIntoDb}
\alias{ctrLoadQueryIntoDb}
\title{Retrieve or update information on clinical trials from register and store in
database}
\usage{
ctrLoadQueryIntoDb(queryterm = "", register = "EUCTR", querytoupdate = 0,
  details = TRUE, parallelretrievals = 10, debug = FALSE,
  collection = "ctrdata", db = "users", url = "mongodb://localhost",
  username = "", password = "", verbose = FALSE)
}
\arguments{
\item{queryterm}{Either a string with the full URL of a search in a register
or the data frame returned by the \link{ctrGetQueryUrlFromBrowser} or the
\link{dbQueryHistory} functions.
The queryterm is recorded in the collection \code{ns} for later use to
update records.}

\item{register}{Vector of abbreviations of registers to query, defaults to
"EUCTR"}

\item{querytoupdate}{Either the word "last" to re-run the last query that
was loaded into the collection, or the integer number of query to be run
again; see \link{dbQueryHistory}. This parameter takes precedence over
\code{queryterm}.}

\item{details}{If \code{TRUE} (default), retrieve full protocol-related
information from EUCTR or XML data from CTGOV, depending on the register
selected. This gives all of the available details for the trials.
Alternatively, set to \code{FALSE} to retrieve only summary information
from EUCTR or CSV data from CTGOV. The full EUCTR information includes
separate records for every country in which the trial is opened; use
function \code{dbFindUniqueEuctrRecord} in a subsequent step to limit to
one record from EUCTR per trial}

\item{parallelretrievals}{Number of parallel downloads of information from
the register}

\item{debug}{Printing additional information if set to \code{TRUE}; default
is \code{FALSE}.}

\item{collection}{Name of collection (default is "ctrdata")}

\item{db}{Name of database (default is "users")}

\item{url}{Address of the mongodb server in mongo connection string URI format
\url{http://docs.mongodb.org/manual/reference/connection-string/} (default is
mongodb://localhost)}

\item{username}{In case access requires credentials.}

\item{password}{In case access requires credentials.}

\item{verbose}{Print information.}
}
\value{
Number of trials imported or updated in the database
}
\description{
Note that upsert is used to store in database, which means that records may
be accumulated. If you want to insert into an empty database, you have to
include a mongo connection object to such an empty database.
}
\examples{
# Retrieve protocol-related information on a single trial identified by EudraCT number
\dontrun{
ctrLoadQueryIntoDb (queryterm = "2013-001291-38")
}

# For use with EudraCT: define paediatric population and cancer terms
\dontrun{
queryEuDefPaedPopulation  <- "age=under-18"
queryEuDef01paedOncTrials <- "cancer leukaem leukem sarcoma tumour tumor blastom gliom lymphom
malign hodgkin ewing rhabdo teratom tumeur leucemi"
queryEuDef01paedOncTrials <- gsub (" ", "\%20OR\%20", queryEuDef01paedOncTrials)
queryEuDef01paedOncTrials <- paste (queryEuDef01paedOncTrials, queryEuDefPaedPopulation, sep="&")
ctrLoadQueryIntoDb (queryterm = queryEuDef01paedOncTrials, parallelretrivals = 5)
}

# Retrieve protocol-related information on ongoing interventional cancer trials in children
\dontrun{
ctrLoadQueryIntoDb (queryterm = "cancer&recr=Open&type=Intr&age=0", register = "CTGOV")
ctrLoadQueryIntoDb (queryterm = "NCT02239861", register = "CTGOV")
}

}
