% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ctrMongo}
\alias{ctrMongo}
\title{This is a function to select and use
a connection to a Mongo DB server.}
\usage{
ctrMongo(mongo = "")
}
\arguments{
\item{mongo}{A mongo data base object, currently using mongolite. Default
is an empty string, which would make this function to result in a default
Mongo DB connection, currently using mongolite with collection ctrdata in
database users on server localhost, unless there is a monoglite data base
object name "mongo" defined in the global environment.}
}
\value{
A mongo data base object, currently using mongolite
}
\description{
This is a function to select and use
a connection to a Mongo DB server.
}
\keyword{internal}

