% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ctrMongo}
\alias{ctrMongo}
\title{Set up connections to a Mongo DB server, one for
the actual trial records and a second for the keys}
\usage{
ctrMongo(collection = "ctrdata", db = "users",
  url = "mongodb://localhost", username = "", password = "",
  verbose = FALSE)
}
\arguments{
\item{collection}{Name of collection (default is "ctrdata")}

\item{db}{Name of database (default is "users")}

\item{url}{Address of the mongodb server in mongo connection string URI format
\url{http://docs.mongodb.org/manual/reference/connection-string/} (default is
mongodb://localhost)}

\item{username}{In case access requires credentials.}

\item{password}{In case access requires credentials.}

\item{verbose}{Print information.}
}
\value{
A mongo data base object, currently using mongolite
}
\description{
Set up connections to a Mongo DB server, one for
the actual trial records and a second for the keys
}
\keyword{internal}
