% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ctrOpenSearchPagesInBrowser}
\alias{ctrOpenSearchPagesInBrowser}
\title{Open advanced search pages of register(s) in default web browser.}
\usage{
ctrOpenSearchPagesInBrowser(register = c("EUCTR", "CTGOV"),
  copyright = FALSE, queryterm = "", ...)
}
\arguments{
\item{register}{Register(s) to open. Either "EUCTR" or "CTGOV" or a vector of both. Default is to open both registers' advanced search pages.}

\item{copyright}{(Optional) If set to \code{TRUE}, opens copyright pages of register.}

\item{queryterm}{(Optional) Show results of search for \code{queryterm} in browser.}

\item{...}{Any additional parameter to use with browseURL, which is called by this function.}
}
\value{
Is always true, invisibly.
}
\description{
Open advanced search pages of register(s) in default web browser.
}

