% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ctrRerunQuery}
\alias{ctrRerunQuery}
\title{ctrRerunQuery}
\usage{
ctrRerunQuery(querytoupdate = querytoupdate,
  forcetoupdate = forcetoupdate, debug = debug,
  collection = collection, uri = uri, password = password,
  verbose = verbose, queryupdateterm = queryupdateterm)
}
\arguments{
\item{querytoupdate}{Either the word "last" or the number of query (based
on \link{dbQueryHistory}) that should be run to retrieve any trial records
that are new or have been updated since this query was run the last time.
This parameter takes precedence over \code{queryterm}.
For EUCTR, updates are available only for the last seven days;
the query is run again if more time has passed since it was run last.}

\item{forcetoupdate}{If \code{TRUE}, run again the query given in
\code{querytoupdate}, irrespective of when it was run last
(default is \code{FALSE}).}

\item{debug}{Printing additional information if set to \code{TRUE}; default
is \code{FALSE}.}

\item{collection}{Name of collection (default is "ctrdata")}

\item{uri}{Default is mongodb://localhost/users/.
Address of database in mongodb server, based on mongo connection string
format: mongodb://[username@]host1[:port1][,...hostN[:portN]]]/database/
Do NOT include password, this will only be used from the parameter.
See \url{http://docs.mongodb.org/manual/reference/connection-string/}}

\item{password}{In case access requires credentials.
Note this defaults to the environment variable "ctrdatamongopassword".
(by means of \code{Sys.getenv("ctrdatamongopassword")}), to
support scripting without revealing secrets.}

\item{verbose}{Print information.}
}
\description{
ctrRerunQuery
}
\keyword{internal}
