% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ctrRerunQuery}
\alias{ctrRerunQuery}
\title{ctrRerunQuery}
\usage{
ctrRerunQuery(querytoupdate = querytoupdate, debug = debug,
  collection = collection, db = db, url = url, username = username,
  password = password, verbose = verbose,
  queryupdateterm = queryupdateterm)
}
\arguments{
\item{querytoupdate}{Either the word "last" to re-run the last query that
was loaded into the collection, or the integer number of query to be run
again; see \link{dbQueryHistory}. This parameter takes precedence over
\code{queryterm}.}

\item{debug}{Printing additional information if set to \code{TRUE}; default
is \code{FALSE}.}

\item{collection}{Name of collection (default is "ctrdata")}

\item{db}{Name of database (default is "users")}

\item{url}{Address of the mongodb server in mongo connection string URI format
\url{http://docs.mongodb.org/manual/reference/connection-string/} (default is
mongodb://localhost)}

\item{username}{In case access requires credentials.}

\item{password}{In case access requires credentials.}

\item{verbose}{Print information.}
}
\description{
ctrRerunQuery
}
\keyword{internal}
