% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dbCTRAnnotateQueryRecords}
\alias{dbCTRAnnotateQueryRecords}
\title{dbQueryAnnotateRecords}
\usage{
dbCTRAnnotateQueryRecords(recordnumbers, annotations, annotation.text,
  annotation.mode, collection = collection, uri = uri,
  password = password, verbose = verbose)
}
\arguments{
\item{annotation.text}{Text to be including in the records retrieved
with the current query, in the field "annotation".}

\item{annotation.mode}{One of "append" (default), "prepend" or "replace"
for new annotation.text with respect to any existing annotation for
the records retreived with the current query.}

\item{collection}{Name of collection (default is "ctrdata")}

\item{uri}{Default is mongodb://localhost/users/.
Address of database in mongodb server, based on mongo connection string
format: mongodb://[username@]host1[:port1][,...hostN[:portN]]]/database/
Do NOT include password, this will only be used from the parameter.
See \url{http://docs.mongodb.org/manual/reference/connection-string/}}

\item{password}{In case access requires credentials.
Note this defaults to the environment variable "ctrdatamongopassword".
(by means of \code{Sys.getenv("ctrdatamongopassword")}), to
support scripting without revealing secrets.}

\item{verbose}{Print information.}
}
\description{
dbQueryAnnotateRecords
}
\keyword{internal}
