% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dbCTRAnnotateQueryRecords}
\alias{dbCTRAnnotateQueryRecords}
\title{dbQueryAnnotateRecords}
\usage{
dbCTRAnnotateQueryRecords(recordnumbers, annotations, annotation.text,
  annotation.mode, con, verbose)
}
\arguments{
\item{annotation.text}{Text to be including in the records retrieved
with the current query, in the field "annotation".}

\item{annotation.mode}{One of "append" (default), "prepend" or "replace"
for new annotation.text with respect to any existing annotation for
the records retreived with the current query.}

\item{con}{A \link[nodbi]{src} connection object, as obtained with
nodbi::\link[nodbi]{src_mongo}() or nodbi::\link[nodbi]{src_sqlite}()}

\item{verbose}{Printing additional information if set to \code{TRUE}; default
is \code{FALSE}.}
}
\description{
dbQueryAnnotateRecords
}
\keyword{internal}
