% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{dbCTRGet}
\alias{dbCTRGet}
\title{Create a data frame from records in the data base that have specified fields in the data base}
\usage{
dbCTRGet(fields = "", mongo = rmongodb::mongo.create(host =
  "127.0.0.1:27017", db = "users"), ns = "ctrdata")
}
\arguments{
\item{fields}{Vector of strings, with names of the sought fields. (Do not use a list.)}

\item{mongo}{(\link{mongo}) A mongo connection object. If not provided, defaults to database "users" on 127.0.0.1 port 27017.}

\item{ns}{Name of the collection in mongo database ("namespace"), defaults to "ctrdata"
useful if the data base includes records from different registers.}
}
\value{
A data frame with columns corresponding to the sought fields. Note that a column for the record _id will always be included.
}
\description{
Create a data frame from records in the data base that have specified fields in the data base
}

