% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dbCTRLoadJSONFiles}
\alias{dbCTRLoadJSONFiles}
\title{dbCTRLoadJSONFiles}
\usage{
dbCTRLoadJSONFiles(dir, con, verbose)
}
\arguments{
\item{dir}{Path to local directory with JSON files
from downloading and converting}

\item{con}{A \link[nodbi]{src} connection object, as obtained with
nodbi::\link[nodbi]{src_mongo}() or nodbi::\link[nodbi]{src_sqlite}()}

\item{verbose}{Printing additional information if set to
\code{TRUE}; default is \code{FALSE}.}
}
\value{
List with elements n (number of imported trials),
_id's of successfully imported trials and
_id's of trials that failed to import
}
\description{
dbCTRLoadJSONFiles
}
\keyword{internal}
