% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dbCTRLoadJSONFiles}
\alias{dbCTRLoadJSONFiles}
\title{dbCTRLoadJSONFiles}
\usage{
dbCTRLoadJSONFiles(dir, mongo)
}
\arguments{
\item{dir}{Path to local directory with JSON files from downloading and converting}

\item{mongo}{ctrmongo Mongo DB database connection object}
}
\value{
List with elements n (number of imported trials) and ids (_ids of imported trials)
}
\description{
dbCTRLoadJSONFiles
}
\keyword{interal}
