% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dbCTRUpdateQueryHistory}
\alias{dbCTRUpdateQueryHistory}
\title{dbCTRUpdateQueryHistory}
\usage{
dbCTRUpdateQueryHistory(register, queryterm, recordnumber, con, verbose)
}
\arguments{
\item{register}{String with abbreviation of register to query,
at this time either "EUCTR" (default) or "CTGOV". Not needed
if \code{queryterm} provide the information which register to
query (see \code{queryterm}).}

\item{queryterm}{Either a string with the full URL of a search in
a register, or the data frame returned by the
\link{ctrGetQueryUrl} or the
\link{dbQueryHistory} functions, or, together with parameter
\code{register}, a string with query elements of a search URL.
The queryterm is recorded in the \code{collection} for later
use to update records.}

\item{con}{A \link[nodbi]{src} connection object, as obtained with
nodbi::\link[nodbi]{src_mongo}() or nodbi::\link[nodbi]{src_sqlite}()}

\item{verbose}{Printing additional information if set to
\code{TRUE}; default is \code{FALSE}.}
}
\description{
dbCTRUpdateQueryHistory
}
\keyword{internal}
