% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbFindIdsUniqueTrials}
\alias{dbFindIdsUniqueTrials}
\title{This function checks for duplicate records of clinical trialss in the
database based on the clinical trial identifier, and it returns a list of ids
of unique trials.}
\usage{
dbFindIdsUniqueTrials(mongo = rmongodb::mongo.create(host = "127.0.0.1:27017",
  db = "users"), ns = "ctrdata", prefermemberstate = "GB",
  include3rdcountrytrials = TRUE, preferregister = "EUCTR")
}
\arguments{
\item{mongo}{(\link{mongo}) A mongo connection object. If not provided,
defaults to database "users" on 127.0.0.1 port 27017.}

\item{ns}{Name of the collection in mongo database ("namespace"), defaults to
"ctrdata"}

\item{prefermemberstate}{Code of single EU Member State for which records should
returned if available. (If not available, a record for GB or lacking this
any other record for the trial will be returned.) For a list of codes of EU
Member States, please see vector \code{countriesEUCTR}.}

\item{include3rdcountrytrials}{A logical value if trials should be retained
that are conducted in third countries, that is outside the European Union.
These can be recognised by EUCTR identifiers ending in -3RD, such as 2010-022945-52-3RD.}

\item{preferregister}{The abbreviation of the preferred register, in case
a trial is in more than one register (string, either "EUCTR" or "CTGOV").
If set to an empty string (""), keeps the keys for the same trial in both
registers in the returned vector.}
}
\value{
A vector with strings of keys (_id in the database) that are
  non-duplicate trials.
}
\description{
If records for a clinical trial are found from more than one register, the
record from EUCTR is returned. The function currently relies on CTGOV
recording other identifiers such as the EudraCT number in the field "Other
IDs".
}
\examples{
\dontrun{
dbFindIdsUniqueTrials()
}

}

