% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbQueryHistory}
\alias{dbQueryHistory}
\title{Show the history of queries that were loaded into a database}
\usage{
dbQueryHistory(collection = "ctrdata", db = "users",
  url = "mongodb://localhost", username = "", password = "",
  verbose = FALSE)
}
\arguments{
\item{collection}{name of collection}

\item{db}{name of database}

\item{url}{address of the mongodb server in mongo connection string
\href{{http://docs.mongodb.org/manual/reference/connection-string/}{URI format}}.}

\item{username}{In case access requires credentials.}

\item{password}{In case access requires credentials.}

\item{verbose}{emit some more output}
}
\value{
A data frame with variables: query-timestamp, query-egister,
 query-records (note: this is the number of records loaded when last executing
 ctrLoadQueryIntoDb(), not the total record number) and query-term
}
\description{
Show the history of queries that were loaded into a database
}
\examples{

\dontrun{
dbQueryHistory()
}

}

