% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbQueryHistory}
\alias{dbQueryHistory}
\title{Show the history of queries that were loaded into a database collection}
\usage{
dbQueryHistory(con, verbose = FALSE)
}
\arguments{
\item{con}{A \link[nodbi]{src} connection object, as obtained with
nodbi::\link[nodbi]{src_mongo}() or nodbi::\link[nodbi]{src_sqlite}()}
}
\value{
A data frame with columns: query-timestamp, query-egister,
 query-records (note: this is the number of records loaded when last executing
 ctrLoadQueryIntoDb(), not the total record number) and query-term,
 and with one row for each ctrLoadQueryIntoDb() loading trial records
 in this collection.
}
\description{
Show the history of queries that were loaded into a database collection
}
\examples{

\dontrun{
dbQueryHistory()
}

}
