% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbQueryHistory}
\alias{dbQueryHistory}
\title{Show the history of queries that were loaded into a database}
\usage{
dbQueryHistory(mongo = "")
}
\arguments{
\item{mongo}{A mongo data base object, currently using mongolite. Default
is an empty string, which would make this function to result in a default
Mongo DB connection, currently using mongolite with collection ctrdata in
database users on server localhost, unless there is a monoglite data base
object name "mongo" defined in the global environment.}
}
\value{
A data frame with variables: query-timestamp, query-egister,
 query-records (note: this is the number of records loaded when last executing
 ctrLoadQueryIntoDb(), not the total record number) and query-term
}
\description{
Show the history of queries that were loaded into a database
}
\examples{
\dontrun{
dbQueryHistory()
}

}

