% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dfFindUniqueEuctrRecord}
\alias{dfFindUniqueEuctrRecord}
\title{Select a single trial record when there are records for different EU Member
States for this trial.}
\usage{
dfFindUniqueEuctrRecord(df = NULL, prefermemberstate = "GB",
  include3rdcountrytrials = TRUE)
}
\arguments{
\item{df}{A data frame created from the database that includes the columns
"_id" and "a2_eudract_number", for example created with function
dbGetVariablesIntoDf(c("_id", "a2_eudract_number")).}

\item{prefermemberstate}{Code of single EU Member State for which records should
returned if available. (If not available, a record for GB or lacking this
any other record for the trial will be returned.) For a list of codes of EU
Member States, please see vector \code{countriesEUCTR}.}

\item{include3rdcountrytrials}{A logical value if trials should be retained
 that are conducted in third countries, that is outside the European Union.
 These can be recognised by EUCTR identifiers ending in -3RD, such as
\href{https://www.clinicaltrialsregister.eu/ctr-search/trial/2010-022945-52/3rd}{2010-022945-52-3RD}.}
}
\value{
A data frame as subset of \code{df} corresponding to the sought
  records.
}
\description{
The EUCTR provides one record per trial per EU Member State in which the
trial is conducted. For all trials conducted in more than one Member State,
this function returns only one record per trial. A preferred Member State can
be specified by the user, and a record of the trial in the preferred Member
State will be returned if available. If not, an english record ("GB") or
lacking this, any other available record will be returned.
}
\details{
Note: To depuplicate trials from different registers (EUCTR and CTGOV),
please first use function \code{\link{dbFindIdsUniqueTrials}}.
}
\keyword{internal}

