% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findBinary}
\alias{findBinary}
\title{Check availability of binaries installed in operating system}
\usage{
findBinary(commandtest = NULL)
}
\arguments{
\item{commandtest}{Command to be used for testing the availability of the binary, e.g. "php -v"}
}
\value{
A logical if executing commandtest returned an error or not
}
\description{
Check availability of binaries installed in operating system
}
\keyword{internal}

