% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{findMongo}
\alias{findMongo}
\title{Convenience function to find location of mongoimport}
\usage{
findMongo()
}
\value{
Either an empty string if \code{mongoimport} was found on the path or, under MS Windows,
a string representing the path to the folder of the mongo binaries
}
\description{
Convenience function to find location of mongoimport
}

