% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{getCTRdata}
\alias{getCTRdata}
\title{Retrieve information on clinical trials from register and store in database}
\usage{
getCTRdata(queryterm = "", register = "EUCTR", updaterecords = FALSE,
  details = TRUE, parallelretrievals = 10,
  mongo = rmongodb::mongo.create(host = "localhost:27017", db = "users"),
  ns = "ctrdata")
}
\arguments{
\item{queryterm}{Part of the URL of a search in a register. The queryterms is recorded in the collection for later use to update records.}

\item{register}{Vector of abbreviations of registers to query, defaults to "EUCTR"}

\item{updaterecords}{Re-run last query for this collection. This parameter takes precedence over \code{queryterm}.}

\item{details}{If \code{TRUE}, retrieve full protocol-related information from EUCTR (rather slow), if \code{FALSE} retrieve only summary information (no country-specific information). Default is \code{TRUE}.}

\item{parallelretrievals}{Number of parallel downloads of information from the register}

\item{mongo}{(\link{mongo}) A mongo connection object. If not provided, defaults to database "users" on localhost port 27017.}

\item{ns}{Name of the collection in mongo database ("namespace"), defaults to "ctrdata"}
}
\value{
Number of trials imported or updated in the database
}
\description{
Note that upsert is used to store in database, which means that records may be accumulated. If you want to insert into an empty database,
you have to include a mongo connection object to such an empty database.
}
\examples{
# Retrieve protocl-related information on a single trial identified by EudraCT number
\dontrun{
getCTRdata (queryterm = "2013-001291-38 ")
}

# For use with EudraCT: define paediatric population and cancer terms, for a query of more than 400 trials
\dontrun{
queryEuDefPaedPopulation  <- "age=adolescent&age=children&age=infant-and-toddler&age=newborn&age=preterm-new-born-infants&age=under-18"
queryEuDef01paedOncTrials <- "cancer leukaem leukem sarcoma tumour tumor blastom gliom lymphom malign hodgkin ewing rhabdo teratom tumeur leucemi"
queryEuDef01paedOncTrials <- gsub (" ", "\%20OR\%20", queryEuDef01paedOncTrials)
queryEuDef01paedOncTrials <- paste (queryEuDef01paedOncTrials, queryEuDefPaedPopulation, sep="&")
getCTRdata (queryterm = queryEuDef01paedOncTrials, parallelretrivals = 5)
}

# Retrieve protocol-related information on ongoing interventional cancer trials in children
\dontrun{
getCTRdata (queryterm = "cancer&recr=Open&type=Intr&age=0", register = "CTGOV")
}
\dontrun{
getCTRdata (queryterm = "NCT02239861", register = "CTGOV")
}
}

