% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{getCTRQueryUrl}
\alias{getCTRQueryUrl}
\title{Import from clipboard the URL of a search in one of the registers}
\usage{
getCTRQueryUrl(content = clipr::read_clip())
}
\arguments{
\item{content}{URL from browser address bar. Defaults to clipboard contents.}
}
\value{
A string of query parameters that can be used to retrieve data from the register.

A list with a query term and the register name that can directly be used in \link{getCTRdata}
}
\description{
Import from clipboard the URL of a search in one of the registers
}
\examples{
\dontrun{
getCTRdata (getCTRQueryUrl())
}
}

