% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{installCygwinWindowsTest}
\alias{installCygwinWindowsTest}
\title{Convenience function to test for working cygwin installation}
\usage{
installCygwinWindowsTest()
}
\value{
Information if cygwin can be used, \code{TRUE} or \code{FALSE}
}
\description{
Convenience function to test for working cygwin installation
}

