% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{installMongoFindBinaries}
\alias{installMongoFindBinaries}
\title{Function to detect the location of mongo database binaries (mongo,
mongoimport) and to save the location into a private environment.
Call this function with a correct parameter if the automatic
detection fails.}
\usage{
installMongoFindBinaries(mongoDirWin = "c:\\\\mongo\\\\bin\\\\")
}
\arguments{
\item{mongoDirWin}{Only used under MS Windows: folder that contains mongo
binaries, defaults to "c:\\mongo\\bin\\" as used on
\url{http://docs.mongodb.org/manual/tutorial/install-mongodb-on-windows#interactive-installation}}
}
\value{
Either an empty string if \code{mongoimport} was found on the path
  or, under MS Windows, a string representing the path to the folder of the
  mongo binaries
}
\description{
Function to detect the location of mongo database binaries (mongo,
mongoimport) and to save the location into a private environment.
Call this function with a correct parameter if the automatic
detection fails.
}
\keyword{internal}

