% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{installMongoFindBinaries}
\alias{installMongoFindBinaries}
\title{Function to detect the location of mongo database binaries (mongo,
mongoimport) and to save the location into a private environment.
Call this function with a correct parameter if the automatic
detection fails.}
\usage{
installMongoFindBinaries(mongoDirWin = "c:\\\\mongodb\\\\bin\\\\",
  mongoDirUnix = "/usr/local/opt/mongodb/", debug = FALSE)
}
\arguments{
\item{mongoDirWin}{Only used under MS Windows: folder that contains mongo
binaries, defaults to "c:\\mongodb\\bin\\". See
\url{http://docs.mongodb.org/manual/tutorial/install-mongodb-on-windows#interactive-installation}}

\item{mongoDirUnix}{Only used under macOS and Linux: folder that contains mongo
binaries, defaults to "/usr/local/opt/mongodb/".}

\item{debug}{Printing additional information if set to \code{TRUE}; default
is \code{FALSE}.}
}
\value{
A vector of length 2 with the paths for the \code{mongo} and
  \code{mongoimport} binaries on the user's system
}
\description{
Function to detect the location of mongo database binaries (mongo,
mongoimport) and to save the location into a private environment.
Call this function with a correct parameter if the automatic
detection fails.
}
\keyword{internal}
